P.H.A.J.

Index

Phaj.Region

Phaj.nn_params

Phaj.build_longest_common_prefix

Phaj.build_suffix_array

Phaj.calculate_thermodynamic_parameters

Phaj.check_probe_tm

Phaj.filter_probes

Phaj.find_homodimers

Phaj.gc_content

Phaj.longest_aligned_region!

Phaj.parse_commandline

Phaj.probe_tm_salt_correction

Phaj.progress_bar_update

Phaj.prpy_sequence

Phaj.sequence_thermodynamic_sum

Phaj.slice_probe

Phaj.sortbyrank

Functions

Documentation for Phaj.jl

Phaj.nn_params — Type

nn_params

A structure encapsulating the thermodynamic parameters for DNA nearest-neighbor (NN) interactions.

The values are based on SantaLucia's 1997 paper titled "Thermodynamics and NMR of internal G·T mismatches in DNA." DOI.

Parameters

The struct contains entropy (ΔS, in cal/mol·K) and enthalpy (ΔH, in kcal/mol) values for each of the NN pairs:

NN Pair	ΔS	ΔH

AA	delta_s	delta_h

AC	delta_s	delta_h

AG	delta_s	delta_h

AT	delta_s	delta_h

CA	delta_s	delta_h

CC	delta_s	delta_h

CG	delta_s	delta_h

GA	delta_s	delta_h

GC	delta_s	delta_h

TA	delta_s	delta_h

Notes

This struct is essential for functions aiming to compute the thermodynamics of DNA sequences or structures.

Phaj.Region — Type

mutable struct Region
    start::Int
    stop::Int
end

A mutable structure that represents a region in a sequence. start and stop are the 1-based indices of the start and end of the region, respectively.

Phaj.progress_bar_update — Function

Update progress bar on the command line.

Phaj.prpy_sequence — Function

prpy_sequence(sequence::AbstractString)::String

Convert a given DNA sequence into purine (A/G represented by 'R') and pyrimidine (C/T represented by 'Y') codes.

Arguments

sequence::AbstractString: A DNA sequence containing nucleotide bases.

Returns

A string where purines are replaced with 'R' and pyrimidines with 'Y'.

Phaj.sequence_thermodynamic_sum — Function

sequence_thermodynamic_sum(delta_h::Vector{Float64}
                            , delta_s::Vector{Float64}
                            , nn_pairs_list::Vector{String}
                            , sequence_nn_list::Vector{String}
                            )

Calculate the total enthalpy and entropy of each sequence.

Arguments

delta_h::Vector{Float64}: A vector containing enthalpy values for nearest-neighbor pairs.

delta_s::Vector{Float64}: A vector containing entropy values for nearest-neighbor pairs.

nn_pairs_list::Vector{String}: A vector containing nearest-neighbor pairs.

sequence_nn_list: A list of nearest-neighbor pairs for the given sequence.

Returns

Total enthalpy and entropy for the sequence.

Phaj.gc_content — Function

gc_content(sequence::LongSequence{DNAAlphabet{4}})::Float64

Calculate the GC content of a given DNA sequence.

Arguments

sequence::LongSequence{DNAAlphabet{4}}: A DNA sequence containing nucleotide bases.

Returns

The percentage of G and C bases in the given sequence.

Phaj.sortbyrank — Function

sortbyrank(i::Int, j::Int, rank::Vector{Int}, k::Int, len::Int) -> Bool

Determine the ordering of two indices in the suffix array based on their ranks.

This function is crucial in the construction of the suffix array. It provides a means to compare two suffixes in the sequence. Initially, the comparison is done based on the individual characters. As the algorithm progresses and the ranks evolve (becoming more refined), the comparison can consider characters further down the sequence.

Arguments

i::Int: The starting index of the first suffix in the sequence.

j::Int: The starting index of the second suffix in the sequence.

rank::Vector{Int}: The rank array corresponding to the current state of suffixes.

k::Int: The offset used to look-ahead in the rank comparisons.

len::Int: The length of the sequence.

Returns

Bool: true if the suffix starting at i comes before the suffix starting at j, false otherwise.

Notes

The function considers not just the current rank but also the rank of the characters k positions ahead in the sequence. This is to ensure the correct ordering of suffixes that may start with the same characters.

Phaj.build_suffix_array — Function

build_suffix_array(sequence::String) -> Vector{Int}

Build a suffix array for a given sequence.

Arguments

sequence::String: A DNA sequence string.

Returns

A Vector{Int} representing the suffix array of the sequence.

Explanation

The suffix array construction algorithm uses a double-sorting technique based on ranks of characters and their next 'k' characters. These ranks are updated in each iteration to reflect the combined ranks of two consecutive substrings of length 'k'.

Phaj.build_longest_common_prefix — Function

build_longest_common_prefix(sequence::String, suffix_array::Vector{Int}) -> Vector{Int}

Build the Longest Common Prefix (LCP) array from a sequence and its suffix array.

Arguments

sequence::String: A DNA sequence string.

suffix_array::Vector{Int}: The suffix array of the sequence.

Returns

A Vector{Int} representing the LCP array of the sequence.

Explanation

The LCP array indicates the number of characters two consecutive suffixes in the suffix array have in common.

Phaj.find_homodimers — Function

find_homodimers(sequence::String) -> Vector{String}

Identify all homodimers in a given DNA sequence using a suffix array and LCP array.

Arguments

sequence::String: A DNA sequence string.

Returns

A Vector{String} containing all identified homodimers in the sequence.

Explanation

This function aims to identify regions in the provided DNA sequence where the sequence is its own complement, known as homodimers. To do this, the sequence is combined with its reverse complement. The reason for using the reverse complement (rather than the simple complement) is that DNA naturally binds in an antiparallel fashion. In other words, when looking for a region in the sequence that can form a homodimer (with itself), it is the reverse complement of a subsequence that will align with and bind to that subsequence.

It then checks the LCP values to identify overlapping regions between the original sequence and its complement, which indicate homodimers.

Phaj.filter_probes — Function

filter_probes(probes::Vector{String}
                , temperature_threshold::Integer
                , monovalent::Float64
                , mg::Float64
                , dntps::Float64
                , oligo_conc::Float64
                , delta_g_threshold::Float64
                , upper_gc::Float64
                , lower_gc::Float64
                , max_aligned_length::Integer
                , max_heterodimer_tm::Integer
                )

Filter probes based on the adjusted melting temperature with respect to a temperature threshold. Calculates the adjusted melting temperature for pairs of sequences and returns those sequences with temperatures below the given threshold. Additionally, calculates and returns Gibbs free energy for the aligned region of each sequence pair.

Arguments

probes::Vector{String}: A vector of DNA sequences (probes) to be filtered.

temperature_threshold::Integer: A temperature threshold (in Celsius) to filter the probes.

monovalent::Int64: Concentration of monovalent ions in mM.

mg::Float64: Concentration of magnesium ions in mM.

dntps::Float64: Concentration of dNTPs in mM.

oligo_c::Float64: Oligo concentration in μM.

delta_g_threshold::Float64: Delta G threshold for homodimer sequences in kcal/mol.

upper_gc::Float64: Upper bound for cutoff based on percent GC, range 0-1.0.

lower_gc::Float64: Lower bound for cutoff based on percent GC, range 0-1.0.

max_aligned_length::Integer: Maximum length of complementary bps a probe can have with other probes.

max_heterodimer_tm::Integer: Maximum melting temperature a stretch of heterodimer bases can have

with another probe.

Returns

promising_probes: A set of sequences with adjusted melting temperatures below the temperature threshold.

Notes

The function considers various thermodynamic and environmental parameters, such as mono- and divalent ion concentrations, to calculate adjusted melting temperatures and Gibbs free energy for each sequence pair.

Phaj.check_probe_tm — Function

check_probe_tm(probe::String
                , temperature_threshold::Integer
                , monovalent::Float64
                , nn::nn_params
                , oligo_c::Float64
                , dntps::Float64
                , mg::Float64
                )

Check if all slices of a DNA probe sequence have melting temperatures (Tm) above a set threshold. The function splits the probe into roughly two equal slices and calculates the Tm for each slice considering corrections for monovalent ions, magnesium ions, and dNTPs.

Arguments

probe::String: The DNA probe sequence to check.

temperature_threshold::Integer: The minimum required melting temperature in degrees Celsius.

monovalent::Int64: Concentration of monovalent ions in mM.

nn::nn_params: A structure containing the thermodynamic parameters for all possible nearest-neighbor pairs.

oligo_c::Float64: Oligo concentration in μM.

dntps::Float64: Concentration of dNTPs in mM.

mg::Float64: Concentration of magnesium ions in mM.

Returns

Boolean: true if all slices have a Tm above the temperature threshold, otherwise false.

Notes

This function aims to ensure that all subsections of a DNA probe sequence have sufficient stability (i.e., high melting temperature) to maintain a double-stranded configuration under specific conditions. It's crucial when designing probes for applications where uniform hybridization across the probe is desired.

Examples

probe = "AATTATGACTGGGAAAGTAAACCGCCTCCACGTAAGCAAGGAAGGCATTCCAATTGTCGAACGGACTGAAGTTTCGGATA"
threshold = 60
monovalent_concentration = 50
nn_parameters = nn_params()
oligo_concentration = 2 * 1e-6
dNTP_concentration = 0
mg_concentration = 2
result = check_probe_tm(probe
                        , threshold
                        , monovalent_concentration
                        , nn_parameters
                        , oligo_concentration
                        , dNTP_concentration
                        , mg_concentration
                        )

Phaj.slice_probe — Function

slice_probe(probe::String, slice_size::Int) -> Vector{String}

Get substring slices from a probe of size slice_size.

Arguments

probe: A string containing the probe sequence to be split.

slice_size: The size of the substring to created. Each probe is divided into chunks of this size (from left to right, and without overlap between chunks).

Returns

Returns the substrings of the input probe sequence.

Example

probe = "CGTGCGCCACTAGACTTGGCAAGGCGTGGAACCGATACCTGCTACCGTGTTAGCAACAAACAGCTATCAACACAGCCATG"
slice_size = 40
probe_slices = slice_probe(probe, slice_size)
2-element Vector{String}:
 "CGTGCGCCACTAGACTTGGCAAGGCGTGGAACCGATACCT"
 "GCTACCGTGTTAGCAACAAACAGCTATCAACACAGCCATG"

Phaj.calculate_thermodynamic_parameters — Function

calculate_thermodynamic_parameters(sequence::AbstractString, monovalent::Float64, nn::nn_params)

Calculate the Gibbs free energy (ΔG), total enthalpy (ΔH), and total entropy (ΔS) of a given DNA sequence. The computation is based on the nearest-neighbor (NN) model, taking into account the sequence's content and a provided monovalent ion concentration.

Arguments

sequence::AbstractString: A DNA sequence for which thermodynamic parameters will be calculated.

monovalent::Int64: The concentration of monovalent ions in mM.

nn::nn_params: A structure containing the thermodynamic parameters for all possible nearest-neighbor pairs.

Returns

aligned_delta_g: The Gibbs free energy (ΔG) of the sequence.

sequence_dh_total: The total enthalpy (ΔH) of the sequence, considering both internal

nearest-neighbor interactions and terminal base pair corrections.

sequence_total_ds: The total entropy (ΔS) of the sequence based on its nearest-neighbor content.

Notes

The function uses nearest-neighbor parameters and end compensation (terminal base pair corrections) to compute the thermodynamics of the DNA sequence. The computed Gibbs free energy is especially important for predicting the stability of DNA duplexes.

Examples

seq = "AGTCGA"
monovalent_concentration = 50
nn_parameters = nn_params()
ΔG, ΔH, ΔS = calculate_thermodynamic_parameters(seq, monovalent_concentration, nn_parameters)

Phaj.longest_aligned_region! — Function

longest_aligned_region!(region::Region, alignment_anchors::Vector{AlignmentAnchor})

Modifies the region object in place to represent the longest aligned region in the given alignment_anchors.

Arguments

region::Region: A Region object that will be modified in place. After the function call,

region.start and region.stop will be the start and end indices of the longest aligned region in the alignment result.

alignment_anchors::Vector{AlignmentAnchor}: The alignment anchors to analyze.

Examples

region = Region(0, 0) 
aln = alignment(alignment_result) 
alignment_anchors = alignment(alignment_result).a.aln.anchors
longest_aligned_region!(region, alignment_anchors)
println("The longest aligned region is from index(region.start) to index(region.stop)") 

Phaj.probe_tm_salt_correction — Function

probe_tm_salt_correction(probe_melting_temperature, monovalent, magnesium, dNTPs, probe_gc, probe_length)

Adjust the provided melting temperature for a DNA probe based on salt and magnesium concentrations in the buffer.

Uses equations from Owczarzy et al. (2008) to modify the melting temperature according to monovalent ion, magnesium, and dNTP concentrations.

Reference: Owczarzy, R., et al. (2008). Biochemistry, https://doi.org/10.1021/bi702363u.

Arguments

probe_melting_temperature::Float64: Initial melting temperature (in Celsius) calculated at 1 M salt concentration.

monovalent::Float64: Concentration of monovalent ions (e.g., Na+ or K+).

magnesium::Float64: Concentration of magnesium ions.

dNTPs::Float64: Concentration of dNTPs.

probe_gc::Float64: GC content of the probe as a fraction (0.0-1.0).

probe_length::Int: Length of the DNA probe.

Returns

Float64: Adjusted melting temperature (in Celsius).

Phaj.parse_commandline — Function

parse_commandline() -> Dict

Parse and collect command line arguments.

Description

This function fetches user-defined parameters for filtering potential DNA hybridization probes. The parameters include monovalent ion concentration, magnesium concentration, dNTP concentration, probe oligo concentration, reaction temperature, homodimer sequence melting temperature, delta G threshold, GC content bounds, maximum length of complementary base pairs, and input/output filenames.

Returns

A dictionary containing user-defined parameters.

Command line arguments:

--mono, -m: Monovalent ion concentration in reaction (default: 50.0 mM).

--mg, -M: Magnesium concentration in reaction (default: 2.0 mM).

--dntps, -d: dNTP concentration in reaction (default: 0.0 mM).

--oligo, -c: Total probe oligo concentration in reaction (default: 0.25 μM).

--temp, -t: Reaction temperature, serving as a melting temperature threshold for probes (default: 65°C).

--heterodimer, -H: Maximum allowed homodimer sequence melting temperature (default: 25°C).

--delta-g, -G: Delta G threshold, below which homodimer sequences are not allowed (default: -10.0).

--upper, -u: Upper bound for percent GC (e.g., 0.7 for 70% GC; default: 0.6).

--lower, -l: Lower bound for percent GC (e.g., 0.45 for 45% GC; default: 0.4).

--max, -L: Maximum length of complementary base pairs a probe can have with other probes (default: 5).

--out, -o: Output filename for a fasta containing promising probes (default: "promising_probes.fa").

fasta: Input fasta file containing hybridization probes to be filtered (required).

Example

# Get arguements from the command-line
parsed_args = parse_commandline()
for (arg,val) in parsed_args
    println("  $arg  =>  $val")
end